requirements.txt
streamlit
pandas
plotly
python-dotenv
koreanize-matplotlib
numpy
